user details added

+++ controllers/books.js
@@ -5,7 +5,7 @@ const { NotFoundError,BadRequestError } = require('../errors')
 const APIFeatures=require("../utils/api-features")
 
 const getAllBooks=async(req,res)=>{
-    const features= new APIFeatures(Book.find(),req.query).paginate()
+    const features= new APIFeatures(Book.find({}),req.query).paginate()
     const book=await features.query
     res.status(StatusCodes.OK).json({ book, count: book.length })
 }
@@ -26,13 +26,17 @@ const createBook=async(req,res)=>{
       );
       await bookImage.mv(imagePath);
       req.body.imageUrl=`/uploads/${bookImage.name}`
+      req.body.createdBy = req.user.userId
       const book=await Book.create(req.body)
       res.status(StatusCodes.CREATED).json({ book })
 }
 
 const getBook=async(req,res)=>{
-    const {id:bookID}=req.params
-        const book=await Book.findOne({_id:bookID})
+    const {
+        user: { userId },
+        params: { id: bookID },
+      } = req
+        const book=await Book.findOne({_id:bookID,createdBy:userId})
         if(!book){
             throw new NotFoundError(`No book with id ${bookID}`)
         }
@@ -40,8 +44,11 @@ const getBook=async(req,res)=>{
 }
 
 const updateBook=async(req,res)=>{
-    const {id:bookID}=req.params
-        const book=await Book.findOneAndUpdate({_id:bookID},req.body,{new:true,runValidators:true})
+    const {
+        user: { userId },
+        params: { id: bookID },
+      } = req
+        const book=await Book.findOneAndUpdate({_id:bookID,createdBy: userId},req.body,{new:true,runValidators:true})
         if(!book){
             throw new NotFoundError(`No book with id ${bookID}`)
         }
@@ -49,8 +56,11 @@ const updateBook=async(req,res)=>{
 }
 
 const deleteBook=async(req,res)=>{
-    const {id:bookID}=req.params
-        const book=await Book.findOneAndDelete({_id:bookID})
+    const {
+        user: { userId },
+        params: { id: bookID },
+      } = req
+        const book=await Book.findOneAndDelete({_id:bookID,createdBy: userId})
         if(!book){
             throw new NotFoundError(`No book with id ${bookID}`)
         }
